# api.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: zhurd
  license:
    name: Apache 2.0
servers:
  - url: http://localhost:3003/v1
paths:
...
  /labels/{labelID}/templates:
    get:
      summary: List all tamplates
      operationId: listTemplates
      tags:
        - templates
      parameters:
        - name: labelID
          in: path
          required: true
          description: The ID of the label
          schema:
            type: string
      responses:
        '200':
          description: An paged array of templates
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Templates'
        '404':
          description: Not found
    post:
      summary: Create a template
      operationId: createTemplate
      tags:
        - templates
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTemplate'
      parameters:
        - name: labelID
          in: path
          required: true
          description: The ID of the label
          schema:
            type: string
      responses:
        '201':
          description: created template
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Template'
        '400':
          description: Invalid request
        '404':
          description: Not found
...
components:
  schemas:
...
    CreateTemplate:
      required:
        - type
        - body
      properties:
        type:
          type: string
          example: ZPL
        body:
          type: string
    Template:
      required:
        - id
        - type
        - body
      properties:
        id:
          type: integer
          format: int64
          example: 1
        label_id:
          type: integer
          format: int64
          example: 1
        type:
          type: string
          example: ZPL
        body:
          type: string
          example: |
            ^XA
            ^FX _comment1_
            ^CF_font1_
            ^FO_indentation_left_top1_^GC_circle_^FS
            ^FO_indentation_left_top2_^GC_circle_^FS
            ^FO_indentation_left_top3_^GC_circle_^FS
            ^FO_indentation_left_top4_^GC_circle_^FS
            ^FO_indentation_left_top5_^GC_circle_^FS
            ^FO_indentation_left_top6_^FD_company_data1_^FS
            ^CF_font2_
            ^FO_indentation_left_top7_^FD_company_data2_^FS
            ^FO_indentation_left_top8_^FD_company_data3_^FS
            ^FO_indentation_left_top9_^FD_company_data4_^FS
            ^FO_indentation_left_top10_^GB_rectangle1_^FS
            ^FX _comment2_
            ^CF_font3_
            ^FO_indentation_left_top11_^FD_data5_^FS
            ^FO_indentation_left_top12_^FD_data6_^FS
            ^FO_indentation_left_top13_^FD_data7_^FS
            ^CF_font4_
            ^FO_indentation_left_top14_^GB_rectangle2_^FS
            ^FX _comment3_
            ^BY_band_
            ^FO_indentation_left_top15_^BC^FD1234567890^FS
            ^FX Fourth section (the two boxes on the bottom).
            ^FO_indentation_left_top16_^GB_rectangle3_^FS
            ^FO_indentation_left_top17_^GB_rectangle4_^FS
            ^CF_font5_
            ^FO_indentation_left_top18_^FD_data8_^FS
            ^FO_indentation_left_top19_^FD_data9_^FS
            ^FO_indentation_left_top20_^FD_data10_^FS
            ^CF_font6_
            ^FO_indentation_left_top21_^FD_data11_^FS
            ^XZ
    Templates:
      type: array
      items:
        $ref: '#/components/schemas/Template'
...